requirements.txt
Django==1.11.9
Werkzeug==0.14.1
coreapi==2.3.3
django-cachalot==1.5.0
django-cleanup==2.1.0
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.6
django-filter==1.1.0
django-jet==1.0.7
django-model-utils==3.1.1
django-redis==4.8.0
django-stdimage==3.0.0
django-taggit-serializer==0.1.5
django-taggit==0.22.2
djangorestframework==3.7.7
djoser==1.1.5
flake8-commas==1.0.0
flake8-isort==2.3
flake8-print==3.0.1
flake8-quotes==0.13.0
flake8==3.5.0
ipdb==0.10.3
ipython==6.2.1
isort==4.2.15
mypy==0.560
pep8-naming==0.5.0
psycopg2==2.7.3.2
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.3.2
pytz==2017.3
raven==6.5.0
redis==2.10.6
requests==2.18.4
social-auth-app-django==2.1.0
social-auth-core[openidconnect]==1.7.0
uWSGI==2.0.15
